Calculates the feature groups importance (called aspects importance) for a selected observation

Aspect Importance function takes a sample from a given dataset and modifies it. Modification is made by replacing part of its aspects by values from the observation. Then function is calculating the difference between the prediction made on modified sample and the original sample. Finally, it measures the impact of aspects on the change of prediction by using the linear model or lasso.

aspect_importance(x, ...)

# S3 method for explainer
aspect_importance(x, new_observation, aspects,
  N = 100, sample_method = "default", n_var = 0, f = 2,
  show_cor = FALSE, ...)

# S3 method for default
aspect_importance(x, data, predict_function = predict,
  new_observation, aspects, N = 100, sample_method = "default",
  n_var = 0, f = 2, show_cor = FALSE, ...)

lime(x, ...)

Arguments

x
a model to be explained or an explainer created with the `DALEX::explain()` function

...
other parameters

new_observation
selected observation with columns that corresponds to variables used in the model

aspects
list containting grouping of features into aspects

N
number of observarions to be sampled from data

sample_method
sampling method in get_sample

n_var
how many non-zero coefficients should be after lasso fitting, if zero than linear regression is used

f
frequency in in get_sample

show_cor
show if all features in aspect are pairwise positivly correlated, works only if dataset contains solely numeric values

data
dataset, it will be extracted from `x`` if it's an explainer

predict_function
predict function, it will be extracted from `x`` if it's an explainer

Value

An object of the class 'aspect_importance'. Contains dataframe that describes aspects' importance.

Examples

library("DALEX")
titanic <- na.omit(titanic)
model_titanic_glm <- glm(survived == "yes" ~
                           class+gender+age+sibsp+parch+fare+embarked,
                         data = titanic, family = "binomial")

explain_titanic_glm <- explain(model_titanic_glm,
                               data = titanic[,-9],
                               y = titanic$survived == "yes")

#> Preparation of a new explainer is initiated
#>   -> model label       :  lm  (default)
#>   -> data              :  2099  rows  8  cols 
#>   -> target variable   :  2099  values 
#>   -> predict function  :  yhat.glm  will be used (default)
#>   -> predicted values  :  numerical, min =  0.009091596 , mean =  0.3244402 , max =  0.9804785  
#>   -> residual function :  difference between y and yhat (default)
#>   -> residuals         :  numerical, min =  -0.9624611 , mean =  -2.959393e-10 , max =  0.967059  
#> A new explainer has been created!

aspects <- list(wealth = c("class", "fare"),
                family = c("sibsp", "parch"),
                personal = c("gender", "age"),
                embarked = "embarked")

aspect_importance(explain_titanic_glm, new_observation = titanic[1,],
                  aspects = aspects)

#>    aspects importance     features
#> 4 personal   -0.18014  gender, age
#> 2   wealth   -0.13168  class, fare
#> 5 embarked   -0.05574     embarked
#> 3   family    0.01448 sibsp, parch

library("randomForest")
model_titanic_rf <- randomForest(survived ~ class + gender + age + sibsp +
                                   parch + fare + embarked, data = titanic)

explain_titanic_rf <- explain(model_titanic_rf,
                              data = titanic[,-9],
                              y = titanic$survived == "yes")

#> Preparation of a new explainer is initiated
#>   -> model label       :  randomForest  (default)
#>   -> data              :  2099  rows  8  cols 
#>   -> target variable   :  2099  values 
#>   -> predict function  :  yhat.randomForest  will be used (default)
#>   -> predicted values  :  numerical, min =  0 , mean =  0.2380772 , max =  1  
#>   -> residual function :  difference between y and yhat (default)
#>   -> residuals         :  numerical, min =  -0.926 , mean =  0.08636303 , max =  1  
#> A new explainer has been created!

aspect_importance(explain_titanic_rf, new_observation = titanic[1,],
                  aspects = aspects)

#>    aspects importance     features
#> 4 personal   -0.22873  gender, age
#> 3   family   -0.04507 sibsp, parch
#> 2   wealth   -0.04009  class, fare
#> 5 embarked   -0.02978     embarked