Calculates Individual Variable Profiles aka Ceteris Paribus Profiles

This explainer works for individual observations. For each observation it calculates Individual Variable Profiles for selected variables. For this reason it is also called 'Local Profile Plot'.

ceteris_paribus(x, ...)

# S3 method for explainer
ceteris_paribus(x, new_observation, y = NULL,
  variables = NULL, variable_splits = NULL, grid_points = 101, ...)

# S3 method for default
ceteris_paribus(x, data, predict_function = predict,
  new_observation, y = NULL, variables = NULL,
  variable_splits = NULL, grid_points = 101, label = class(x)[1],
  ...)

Arguments

x
a model to be explained, or an explainer created with function `DALEX::explain()`.

...
other parameters

new_observation
a new observation with columns that corresponds to variables used in the model

y
true labels for `new_observation`. If specified then will be added to ceteris paribus plots.

variables
names of variables for which profiles shall be calculated. Will be passed to `calculate_variable_splits()`. If NULL then all variables from the validation data will be used.

variable_splits
named list of splits for variables, in most cases created with `calculate_variable_splits()`. If NULL then it will be calculated based on validation data avaliable in the `explainer`.

grid_points
number of points for profile. Will be passed to `calculate_variable_splits()`.

data
validation dataset, will be extracted from `x` if it's an explainer

predict_function
predict function, will be extracted from `x` if it's an explainer

label
name of the model. By default it's extracted from the 'class' attribute of the model

Value

An object of the class 'ceteris_paribus_explainer'. It's a data frame with calculated average responses.

Examples

library("DALEX")

library("titanic")
library("randomForest")

titanic_small <- titanic_train[,c("Survived", "Pclass", "Sex", "Age",
                                   "SibSp", "Parch", "Fare", "Embarked")]
titanic_small$Survived <- factor(titanic_small$Survived)
titanic_small$Sex <- factor(titanic_small$Sex)
titanic_small$Embarked <- factor(titanic_small$Embarked)
titanic_small <- na.omit(titanic_small)
rf_model <- randomForest(Survived ~ Pclass + Sex + Age + SibSp + Parch + Fare + Embarked,
                         data = titanic_small)
predict_fuction <- function(m,x) predict(m, x, type="prob")[,2]
explainer_rf <- explain(rf_model, data = titanic_small,
                        y = titanic_small$Survived == "1", label = "RF",
                        predict_function = predict_rf_fuction)

#> Error in explain(rf_model, data = titanic_small, y = titanic_small$Survived ==     "1", label = "RF", predict_function = predict_rf_fuction): object 'predict_rf_fuction' not found

selected_passangers <- select_sample(titanic_small, n = 100)
cp_rf <- ceteris_paribus(explainer_rf, selected_passangers)

#> Error in class(profile) <- c("individual_variable_profile", class(profile)): attempt to set an attribute on NULL

cp_rf

#> Error in eval(expr, envir, enclos): object 'cp_rf' not found

plot(cp_rf, variables = "Age", color = "grey") +
show_observations(cp_rf, variables = "Age", color = "grey") +
  show_rugs(cp_rf, variables = "Age", color = "red")

#> Error in plot(cp_rf, variables = "Age", color = "grey"): object 'cp_rf' not found