Function for getting binary matrix

Function creates binary matrix. It starts with a zero matrix. Then it replaces some zeros with ones. It either randomly replaces one or two zeros per row. Or replace random number of zeros per row - average number of replaced zeros can be controled by parameter f. Function doesn't allow the returned matrix to have rows with only zeros.

get_sample(n, p, method = c("default", "binom"), f = 2)

Arguments

n
number of rows

p
number of columns

method
sampling method

f
probability for binomial sampling

Value

a binary matrix

Examples

 

 get_sample(100,6,"binom",3)

#>        [,1] [,2] [,3] [,4] [,5] [,6]
#>   [1,]    0    0    0    0    0    1
#>   [2,]    0    0    0    0    1    1
#>   [3,]    0    0    1    0    0    0
#>   [4,]    0    0    1    1    1    0
#>   [5,]    0    1    0    0    0    0
#>   [6,]    1    0    0    0    0    0
#>   [7,]    1    0    0    0    1    1
#>   [8,]    1    0    1    0    0    0
#>   [9,]    0    0    1    0    1    1
#>  [10,]    0    0    1    0    0    0
#>  [11,]    0    1    1    1    1    0
#>  [12,]    1    0    0    0    1    0
#>  [13,]    0    1    0    0    1    0
#>  [14,]    1    0    1    1    0    1
#>  [15,]    0    1    1    1    1    0
#>  [16,]    1    1    0    0    0    0
#>  [17,]    0    0    1    0    1    0
#>  [18,]    1    0    0    1    0    1
#>  [19,]    0    0    0    0    1    1
#>  [20,]    0    1    0    1    0    1
#>  [21,]    0    1    0    1    1    1
#>  [22,]    1    0    1    0    0    1
#>  [23,]    1    1    0    0    1    0
#>  [24,]    1    1    0    1    1    0
#>  [25,]    0    1    0    1    0    0
#>  [26,]    0    0    1    0    0    0
#>  [27,]    1    1    0    0    0    0
#>  [28,]    1    1    0    0    0    0
#>  [29,]    1    0    1    1    0    0
#>  [30,]    0    0    1    0    1    1
#>  [31,]    0    0    0    1    0    0
#>  [32,]    0    1    1    0    0    1
#>  [33,]    0    0    0    1    1    0
#>  [34,]    1    0    0    1    1    1
#>  [35,]    0    0    0    1    1    1
#>  [36,]    0    0    0    0    0    1
#>  [37,]    1    1    1    0    0    0
#>  [38,]    0    0    1    1    0    1
#>  [39,]    0    0    1    0    0    0
#>  [40,]    0    1    1    0    1    0
#>  [41,]    1    0    0    0    0    1
#>  [42,]    0    0    0    0    0    1
#>  [43,]    0    1    1    0    0    1
#>  [44,]    0    1    1    0    0    0
#>  [45,]    0    1    1    1    1    0
#>  [46,]    0    1    0    0    0    0
#>  [47,]    1    1    0    0    1    1
#>  [48,]    0    1    0    1    1    0
#>  [49,]    1    0    0    1    0    0
#>  [50,]    1    0    0    0    0    0
#>  [51,]    0    1    0    0    0    1
#>  [52,]    0    1    0    0    0    1
#>  [53,]    1    0    0    0    0    1
#>  [54,]    0    0    1    0    0    1
#>  [55,]    0    0    1    1    1    1
#>  [56,]    0    0    0    1    1    0
#>  [57,]    0    0    1    0    1    0
#>  [58,]    1    0    0    1    0    1
#>  [59,]    1    0    1    1    1    0
#>  [60,]    0    1    1    1    0    0
#>  [61,]    0    1    1    0    1    1
#>  [62,]    0    1    0    0    0    0
#>  [63,]    0    1    0    0    0    0
#>  [64,]    1    1    1    0    1    0
#>  [65,]    1    0    0    0    1    0
#>  [66,]    1    0    0    0    0    0
#>  [67,]    1    0    1    0    0    1
#>  [68,]    0    1    1    1    0    0
#>  [69,]    0    1    0    0    1    1
#>  [70,]    0    1    0    1    0    0
#>  [71,]    1    1    0    1    0    0
#>  [72,]    0    0    0    1    0    1
#>  [73,]    0    1    0    0    0    1
#>  [74,]    1    0    0    0    1    0
#>  [75,]    0    0    1    0    0    1
#>  [76,]    0    0    0    0    0    1
#>  [77,]    0    1    1    0    1    1
#>  [78,]    0    0    1    1    0    1
#>  [79,]    0    0    0    0    0    1
#>  [80,]    0    0    1    0    0    1
#>  [81,]    1    0    0    1    1    0
#>  [82,]    0    0    0    1    0    0
#>  [83,]    0    1    0    0    0    1
#>  [84,]    1    0    0    0    0    1
#>  [85,]    1    0    0    0    0    0
#>  [86,]    1    0    0    0    0    0
#>  [87,]    0    1    0    1    1    1
#>  [88,]    1    0    0    1    0    0
#>  [89,]    1    1    0    1    0    1
#>  [90,]    0    1    0    1    0    0
#>  [91,]    0    0    0    1    1    0
#>  [92,]    1    1    1    0    1    1
#>  [93,]    1    0    0    0    1    0
#>  [94,]    1    0    1    0    1    1
#>  [95,]    0    1    0    0    1    1
#>  [96,]    0    0    1    0    1    0
#>  [97,]    0    1    1    1    0    1
#>  [98,]    0    1    0    0    0    1
#>  [99,]    0    0    1    0    0    1
#> [100,]    0    1    0    1    0    1