Plots Ceteris Paribus Profiles in D3 with r2d3 Package.

Function 'plotD3.ceteris_paribus_explainer' plots Individual Variable Profiles for selected observations. Various parameters help to decide what should be plotted, profiles, aggregated profiles, points or rugs.

Find more detailes in Ceteris Paribus Chapter.

plotD3(x, ...)

# S3 method for ceteris_paribus_explainer
plotD3(x, ..., size = 2, alpha = 1,
  color = "#46bac2", only_numerical = TRUE, facet_ncol = 2,
  scale_plot = FALSE, variables = NULL, chartTitle = NULL,
  label_margin = 60, show_observations = TRUE, show_rugs = TRUE)

Arguments

x
a ceteris paribus explainer produced with function `ceteris_paribus()`

...
other explainers that shall be plotted together

size
a numeric. Set width of lines

alpha
a numeric between 0 and 1. Opacity of lines

color
a character. Set line color

only_numerical
a logical. If TRUE then only numerical variables will be plotted. If FALSE then only selected variables will be plotted as bars.

facet_ncol
number of columns for the `facet_wrap`

scale_plot
a logical. Should size of the plot scale with window size? By default it's FALSE

variables
if not NULL then only `variables` will be presented

chartTitle
a character. Set custom title

label_margin
a numeric. Set width of label margins, when only_numerical is FALSE

show_observations
a logcal. Adds observations layer to a plot. By default it's TRUE

show_rugs
a logcal. Adds rugs layer to a plot. By default it's TRUE

Value

an `r2d3` object.

Examples

# NOT RUN {
library("DALEX")
library("ingredients")
library("randomForest")

titanic <- na.omit(titanic)

model_titanic_rf <- randomForest(survived == "yes" ~ gender + age + class + embarked +
                                 fare + sibsp + parch,  data = titanic)
explain_titanic_rf <- explain(model_titanic_rf,
                             data = titanic[,-9],
                             y = titanic$survived == "yes",
                             label = "rf")

selected_passangers <- select_sample(titanic, n = 10)
cp_rf <- ceteris_paribus(explain_titanic_rf, selected_passangers)

plotD3(cp_rf, variables = c("age","parch","fare","sibsp"),
     facet_ncol = 2, scale_plot = TRUE)

selected_passanger <- select_sample(titanic, n = 1)
cp_rf <- ceteris_paribus(explain_titanic_rf, selected_passanger)

plotD3(cp_rf, variables = c("class", "embarked", "gender", "sibsp"),
     facet_ncol = 2, only_numerical = FALSE, label_margin = 100, scale_plot = TRUE)
# }