Aspect Importance This explainer calculates the feature groups importance (called aspects importance) for a selected observation.

Aspect Importance function takes a sample from a given dataset and modifies it. Modification is made by replacing part of its aspects by values from the observation. Then function is calculating the difference between the prediction made on modified sample and the original sample. Finally, it measures the impact of aspects on the change of prediction by using the linear model or lasso.

aspect_importance(x, ...)

# S3 method for explainer
aspect_importance(x, new_observation, aspects_list,
  B = 100, method = "default", n_var = 0, ...)

# S3 method for default
aspect_importance(x, data, predict_function = predict,
  new_observation, aspects_list, B = 100, method = "default",
  n_var = 0, ...)

lime(x, ...)

Arguments

x
a model to be explained or an explainer created with the `DALEX::explain()` function

...
other parameters

new_observation
selected observation with columns that corresponds to variables used in the model

aspects_list
list containg grouping of features into aspects

B
number of rows to be sampled from data

method
sampling method in get_sample()

n_var
how many non-zero coefficients for lasso fitting, if zero than linear regression is used

data
dataset, it will be extracted from `x`` if it's an explainer

predict_function
predict function, it will be extracted from `x`` if it's an explainer

Value

An object of the class 'aspect_importance'. Contains dataframe that describes aspects' importance.

Examples

library("DALEX")
titanic <- na.omit(titanic)
model_titanic_glm <- glm(survived == "yes" ~
                           class+gender+age+sibsp+parch+fare+embarked,
                         data = titanic, family = "binomial")

explain_titanic_glm <- explain(model_titanic_glm,
                               data = titanic[,-9],
                               y = titanic$survived == "yes")

aspects <- list(wealth = c("class", "fare"),
                family = c("gender", "sibsp", "parch"),
                age = "age",
                embarked = "embarked")

aspect_importance(explain_titanic_glm, new_observation = titanic[1,],
                  aspects_list = aspects)

#>    aspects  importance
#> 2   wealth -0.19077283
#> 3   family -0.16457823
#> 4      age -0.12725353
#> 5 embarked -0.03651311

library("randomForest")
model_titanic_rf <- randomForest(survived ~ class + gender + age + sibsp +
                                   parch + fare + embarked, data = titanic)

explain_titanic_rf <- explain(model_titanic_rf,
                              data = titanic[,-9],
                              y = titanic$survived == "yes")

aspect_importance(explain_titanic_rf, new_observation = titanic[1,],
                  aspects_list = aspects)

#>    aspects  importance
#> 3   family -0.11604407
#> 4      age -0.09700147
#> 2   wealth -0.08162348
#> 5 embarked  0.01873581