Select Subset of Rows Closest to a Specified Observation

This function selects subset of rows from data set. This is useful if data is large and we need just a sample to calculate profiles.

select_neighbours(observation, data, variables = NULL,
  distance = gower::gower_dist, n = 20, frac = NULL)

Arguments

observation
single observation

data
set of observations

variables
names of variables that shall be used for calculation of distance. By default these are all variables present in `data` and `observation`

distance
the distance function, by default the `gower_dist` function.

n
number of neighbours to select

frac
if `n` is not specified (NULL), then will be calculated as `frac` * number of rows in `data`. Either `n` or `frac` need to be specified.

Value

a data frame with selected rows

Details

Note that select_neighbours function is S3 generic. If you want to work on non standard data sources (like H2O ddf, external databases) you should overload it.

Examples

library("DALEX")

new_apartment <- apartments[1, 2:6]
small_apartments <- select_neighbours(apartmentsTest, new_apartment, n = 10)
new_apartment

#>   construction.year surface floor no.rooms    district
#> 1              1953      25     3        1 Srodmiescie

small_apartments

#>      construction.year surface floor no.rooms district
#> NA                  NA      NA    NA       NA     <NA>
#> NA.1                NA      NA    NA       NA     <NA>
#> NA.2                NA      NA    NA       NA     <NA>
#> NA.3                NA      NA    NA       NA     <NA>
#> NA.4                NA      NA    NA       NA     <NA>
#> NA.5                NA      NA    NA       NA     <NA>
#> NA.6                NA      NA    NA       NA     <NA>
#> NA.7                NA      NA    NA       NA     <NA>
#> NA.8                NA      NA    NA       NA     <NA>
#> NA.9                NA      NA    NA       NA     <NA>